+++ models/__init__.py
@@ -1,3 +1,5 @@
 from . import elit_config
+from . import elit_catalog_line
+from . import elit_catalog_run
 from . import product_template
 from . import sync_processor