+++ tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_elit_sync
+from . import test_elit_catalog